+++ lib/features/cart/repositories/cart_repo.dart
@@ -41,7 +41,7 @@ class CartRepo {
   static Future<dynamic> addOrUpdateCart({
     required String pCode,
     required String iCode,
-    required int qty,
+    required double qty,
     required double rate,
   }) async {
     String? token = await SecureStorageHelper.read(
