+++ lib/features/products/repositories/products_repo.dart
@@ -59,7 +59,7 @@ class ProductsRepo {
   static Future<dynamic> addOrUpdateCart({
     required String pCode,
     required String iCode,
-    required int qty,
+    required double qty,
     required double rate,
   }) async {
     String? token = await SecureStorageHelper.read(
